// frontend/src/sellerSection/lib/types/sellerProductTypes.ts
export type sellerProducts = {
    product_id: string;        
    product_name: string; 
    product_description: string;
    category: string;
    approval_status: string;   
    created_at: Date;
    updated_at: Date;   
    stock: number;   
    seller_id: string;      
    price: number;  
    category_type: string;      
}